Issue #4689: Disallow simultaneous segmentations
The 'start segmentation' button is disabled when a segmentation is already
being worked on. Also, if you change the selected image in the data manager
e.g. to another reference image or segmentation, it will not affect the
segmentors. The selected images will not change in them.

Until you discard or finalise the editing (Cancel or OK), you cannot select
another image to edit. More precisely, you can select another image but it
will not interfere with the current segmentation.

This is not 'disabling' a feature, the volume editors were not designed with
allowing simultaneous editing in mind. However, this restriction was not forced
on the GUI. This commit makes the GUI behave consistently with the design.