Issue #4675: InteractionEventObserverMutex class to prevent interfering interactions
A new class 'InteractionEventObserverMutex' has been introduced to
prevent processing multiple interactions by different state machines at
the same time.

This can happen for example if an interactor changes state when a mouse
button is pressed and goes back to the start state when the button is
released. If another interactor is active at the same time, it can
process key events, while the mouse button is pressed and the first
interactor is doing something.

In case of the general segmentor draw tool and 'ToolKeyPressStateMachine',
one can press 'c' to clean a slice while s/he is erasing contour with
the middle mouse button pressed. This should not be allowed, really.

When the mutex is instantiated, it saves the actual state of the
currently registered interactors (whether they are enabled or disabled),
and disables them. When the mutex dies, it restores the original state
of the interactors.