Issue #4704: Remove duplicated code from commonapps plugin
The code was a duplication of classes in org.mitk.gui.qt.ext so that
we can remove the dependency on that plugin. The problem with that
plugin was that it processed the command line arguments, and it was
not possible to introduce new options in our app (NiftyMIDAS,
specifically).

The new idea is that we introduce a flag 'applicationArgs.processByMITK'
that is checked in org.mitk.gui.qt.ext and it processes the arguments
only if the flag is not set or it evaluates to 'true'.

The commonapps plugin activator also checks the flag and processes the
arguments if it is set to 'false'. The flag should be set to 'false' for
NiftyMIDAS, and possibly for NiftyView and NiftyIGI as well, as the
options are for generic purposes, e.g. setting data node properties or
parent-child relationships between nodes, that is valid for any
applications.