Issue #4747: DnDDisplay module unit test fixes
Some of these have been fixed on the MITK upgrade branch already,
and are ported back now.

Some others are workarounds because of the geometry initialisation
bug. A constant has been introduced that tells if the patch for
the bug is applied to the MITK sources. The value of the constant
must be set manually.