Issue #4750: Allow setting lookup table by command line option
The lookup table can now be set like this:

  --property <data node>:LookupTableName=<lookup table name>

The commit has some minor fixes as well, like eventual overindexing with incorrect
command line options (see the 'resize()' call in the commit), or adding more check
for right use of command line options.

The commit includes adding a listener to the change of the LookupTableName property.
When this property changes, the lookup table with the given name is set for the
data node.