Issue #3344: Make DataStorageListener and DataNodePropertyListener constructors public
This is necessary so that you can instantiate the classes directly and
use them by assigning handlers to their mitk::MessageN typed data members.

This makes the class a bit easier to reuse, as you do not need to derive
a class every time if you want to listen to these events.