Issue #4597: Remove 'Build Utilities' option
The option enabled to build one utility application that fixes the header includes
if they are not in the standard format. The app was written in C++ and it did not
compile because of a missing dependency. Although this could have been fixed, the
app has not been used for a while, and this kind of utilities should be written
in Python or Bash rather than C++.