Issue #3927: Display place holder annotations and control through GUI
The command line option value does not come through yet.

Also, it has to be decided if the property annotations should always be shown
in every viewer or only in specific viewer. The current command line option
processing allows the later, but if there is no "annotation binding" then
it can be annoying to set the properties for annotation for each viewer one
by one.