Issue #3927: Replace ctkCheckablePushButton by a checkbox and a custom label
The custom niftk::ClickableLabel class derives from QLabel and sends 'clicked()'
signals on mouse clicked events.

The label is better solution than the clickable push button for two reasons:

First, there was a weird issue on Mac. If you clicked on the push button, the
check box jumped a bit left and up. When the action finished, it went back to
its place.

Second, the buttons are not transparent by default but labels are. Since these
controls are on a transparent widget, it matters.