Issue #4761: Install NifTK for managed machine from cmiclab-lin
The installation was done on a managed machine, but the packaging
always stopped after an hour or so. With this change, the package
created on another machine (cmiclab-lin now) is installed on the
managad machines (shared drive).

Note that passwordless authentication must be set to ssh from
gitlab-runner@cmiclab-lin to logno, and the user must have write
access to the target directory.