Issue #4776: Visibility change observer moved to intensity profile plugin
The class is obsolete, the plugin is also obsolete. The class has been
moved to the obsolete plugin so that people do not start using it.

The functionality is provided by the property listener classes in niftkCore.

Ideally, the intensity profile plugin should be changed to use this
property listener, but since the plugin is obsolete and switched off
by default, it is very low priority.