Issue #4682: gflags and glog to lower case.
The reason for this change is that gflags and glogs do provide
gflags-config.cmake and gloga-config.cmake and so in theory it
should be possible to add these packages to CMAKE_PREFIX_PATH
and discover them in CMakes Config mode. However, they
currently don't really work, so I had to revert back to CMake's
Module mode. Maybe in the future we can use Config mode, where
the case of the project has to match the name of the config file
which in this instance should be lower case.