Issue #4778: Make 3D test image with check pattern
This commit changes the previous behaviour when generating chessboard
calibration images. These images (mode 6 and 7) were always saved as
2D.

With the current commit, check pattern images are saved as 3D by default,
like any other images. If you need to drop one dimension, you can use
the new '-collapse <axis>' option.

Note that in mode 7 the chess board is in the middle of the volume,
that is if the volume is wider then the chess board then the first
slice will not have a check pattern in the middle. Therefore, in mode
7 you need to pass the '-nz 1 -collapse 2' options to get the previous
behaviour.