Issue #4783: Display position and intensity annotations in side viewer
This is for testing only. If not this bug, the position and the intensity should
be the same as in the main viewer (DnD Display), so displaying them in the side
viewer as well would be redundant.

This commit will have to be reverted after the bug will have been fixed.