Issue #4783: Update position and intensity annotations when position changes
The annotations were updated if the position was changed via the display interactor
or when the window layout or the selected window has changed, but not when
the selected position was changed by other means, e.g. clicking in the side viewer.

(The slice navigation controllers of these windows are connected.)