Issue #4783: Remove 0.5 from displayed index positions for non-image geometries
The cursor (aka. crosshair) is always put in voxel centre that is we can select
only voxels, not arbitrary mm position within a voxel. Therefore, the displayed
voxel indices should be integer numbers. However, for non-image geometries the
origin is in the bottom-left-back corner of the volume, not in voxel centre but
in the corner point of the voxel in the corner. So, the continuous indices of the
cursor will always end in '.5'.

The main viewer is always initialised based on an image geometry, but the side
viewer is initialised based on the main viewer that has non-image geometry. So
that they show the same indices, we have to subtract 0.5 from the coordinates
for non-image geometries.

Note that the position annotation is disabled for the side viewer, as it does
not provide any additional information. (You cannot select a voxel different
than in the main viewer.) However, fixing this can allow initialising the main
viewer (DnD Display) from non-image geometries, e.g. surfaces, later.