Issue #2525: Exclude DnDDisplay continuous tests on Mac
The changes have been tested on Mac OS X 10.11 and all the tests pass.
This commit excludes the tests only in the continuous test script, but
it does not disable them in CMake. The tests would fail on the Mac runner
because the gitlab-runner user does not have access to the display and
the display redirection does not work.