Issue #4806: Do not listen to data storage events in thumbnail viewer
The thumbnail viewer should use the same rendering manager as the tracked
renderer, and it does so. Therefore, it is updated together with the tracked
renderer. Therefore, is no need to listen to data store events (node addition
and removal and data node property changes.)

Instead, it is enough to listen to changes in the focused renderer (which
renderer to track) and its world geometry, selected slice and display geometry.