Issue #4806: Disable mitk::DisplayInteractor when thumbnail interactor is active
When the thumbnail viewer view is activated, the thumbnail interactor is
enabled and every other interactor is disabled. When the view gets deactivated,
the thumbnail interactor is disabled and the other interactors which were
enabled when the thumbnail viewer got active, are switched back on.

This commit also adds an alternative configuration for panning the image
via the thumbnail viewer. You could do it with the middle mouse button,
now shift+left mouse button works as well. This is to make it similar to
the DnD Display interactions.