Issue #4808: Removing 'keep image directions' feature
When the 'keep image directions' and/or 'keep non-image directions' flags were
set in the default world direction constant, the renderer geometries were created
in the way that the renderer plane normals pointed in the same direction as in
the original, reference geometry. With other words, the slices were numbered
in the same order as in the reference geometry, in each renderer.

However, this does not allow having different default world directions in NifTK
than in MITK. In the data manager you can do 'Reinit' or 'Global reinit', and in
both case, the axial slices will be numbered from top to bottom in the axial
renderer. With 'Reinit' the reference geometry will be an image geometry, with
'Global reinit' it will be a non-image geometry.

If the DnD Display has different initialisation strategy for 'image' reference
geometries than MITK does, the 'Reinit' will not work.

Therefore, the initialisation strategy *must* be the same for both image and
non-image geometries.

The current commit sets the MITK initialisation mode, but leaves a constant
(RENDERER_DIRECTIONS) so that it can be changed later easily, if the pull
request for the MITK bug 20180 gets accepted.

Since the current renderer directions do not follow neither the image directions
nor the world directions, you must *not* rely on the indices in the slice
navigation controller of the renderers, but you always need to translate world
coordinates to index coordinates using the reference geometry that was used to
initialise the viewer.