Issue #4808: MITK update to revert fixes for sliced geometry initialisation
I had to give up on correcting the renderers' geometry. With a previous
MITK patch (see MITK bug T20204) you can access the reference geometry
of the renderers' geometry, i.e. the geometry that was used to initialise
the renderers.

The renderers' geometry should not be used for any calculation. You should
always get their reference geometry and use that one for checking corners,
converting between index and mm coordinates and so on.