Issue #4832: nullptr check for renderer geometry when updating position annotation
When the viewer is being initialised with a new geometry, Modified signals are
sent out from the renderers, and at the end of a complicated series of events,
updating the position annotation is requested. Updating the position annotation
requires the world geometry of the selected renderer, in debug mode only.

However, the signals are sent out when the first renderer gets the geometry,
and the selected renderer (if not the first) will not have a valid geometry yet.