Issue #4808: Imitating how MITK creates geometry for renderers
Although the way how MITK constructs geometries for the renderers is kind of arbitrary
and inconsistent, we will imitate the same way for three reasons.

1.

The renderer geometries have weird coordinate system, therefore you should never rely
on the renderer geometries when doing any calculation. Instead, you should get the
'reference' geometry of the geometry of the renderers, i.e. the geometry that was used
to initialise the renderer. However, if you do not use the geometry of the renderers,
it is indifferent in which way they are constructed, until what is rendered is correct.
The DnD Display and the side viewer always uses the reference geometry. See below for
more details which parameters of the renderer geometry are 'wrong'. Of course, if you
know in which way they are 'wrong’, you can compensate this and calculate the corrected
coordinates.

1.a.

The geometry of the renderer is shifted with half a voxel along to the renderer plane
axis, compared to the reference geometry. This is done so that the crosshair planes
are rendered in the correct position, i.e. in the middle of the slices. (The current
mappers are written in this way.)

Because of the half voxel shift, the origin and the corner points of the renderer
geometries do not match those of the reference geometry. Therefore, you *must not* use
the renderer geometry to translate world coordinates to voxel indices.

1.b.

The renderer geometries are sliced geometries that are composed of plane geometries.
Plane geometries are not actual 2D geometries but 3D geometries of one voxel thickness.
MITK does not allow creating plane geometries with left handed coordinate system. If
you try to do so (supplying an inverted normal or negative thickness), the normal will
be overwritten by the cross-product of the right and bottom vectors, and will therefore
be always right-handed.

Sliced geometries are a stack of plane geometries. When creating a sliced geometries,
you can choose in which way you want to 'stack them up'. If you specify the 'flipped'
property, you stack the planes up in opposite direction to their normal. It causes the
weird situation that the plane normals will point opposite to the sliced geometry
normal, and the bounding box of the first plane (remember of the one voxel thickness)
will be outside of the volume of the sliced geometry. Anyhow, this looks very bad.

You *must not* rely on any of these information:

 - corner points of the plane geometries composing the renderer geometry
 - centre of the plane geometries
 - normal of the plane geometries

You can safely rely on the origin and the right and bottom vectors of the plane
geometries, though.

1.c.

The first two coordinates of the renderer geometries have to be fixed so that the
image is rendered in the expected way for each orientation. The third coordinate,
the one that is orthogonal to the renderer plane, can be chosen. This is, basically,
the direction how we 'stack up' the plane geometries when we construct the sliced
geometry for the renderers. In principle, the handedness should be the same in each
renderer, so that the volume can be translated from one renderer geometry to the
other by only rotation, but without mirroring.

However, the sagittal and axial renderers have right handed coordinate system, but
the coronal renderer has left handed one.

1.d.

The direction of the plane axis is inverted for the axial renderer. This causes that
handedness of the world coordinate system is kept. (For the axial renderer the coronal
direction is inverted as well, so that the image is displayed according to the
'eyes-up' convention. Flipping any two coordinates preserves the handedness.)

However, this inverts the slice indexing in the axial renderer compared to the slice
indexing in the world coordinate system.

For this reason, you *must not* rely on slice indexes in the renderer geometries,
and must not use the renderer geometries to translate between world coordinates
and voxel coordinates.

Note that it might make sense to flip the axial direction to preserve handedness
in the axial renderer. However, it should be done for the coronal renderer as well
then. But it is not.

2.

MITK display compatibility

Changing how geometries are constructed in the DnD Display would brake the side viewer
when used with the MITK Display.

The side viewer catches focus change events and initialises its 2D render windows
based on the reference geometry of the focused renderer. Its purpose is to show the
same world as in the main display but from a different orientation.

When the focus goes to a new render window, the slice navigator controllers of the
two viewers are connected. Axial to axial, sagittal to sagittal, coronal to coronal.
The connection is direct, the selected index is not translated. If a position is
selected in one display, its slice navigation controllers send out events with the
new index (!) position, and the position with the same index (!) is selected in
the other display.

So that the same indices mean the same world position in mm, the renderer geometries
must be constructed with the same plane directions.

Other solution would be to translate the indices but this would complicate the logic
significantly.

3.

Image navigator compatibility

The image navigator view connects sliders to the slice navigation controllers of the
renderers of the main display. The connection is not direct as between the slice
navigation controllers of the side viewer and the main display, but it is done through
'adapters’. These adapters have an option to invert the index before delegating it
from the slider to the slice navigation controller or back.

It is now hard-coded in the image navigator, for which orientations it inverts the
direction. It is inverted only for the axial orientation. This matches how geometries
are constructed in MITK. However, if we would construct the geometries in a different
way then the image navigator would display incorrect (flipped) indices.

Note that the image navigator shows indices in world space, always. This is fine if
the renderers are initialised using a non-image geometry describing a world space.
However, if they were initialised using an image geometry, you would expect to see
and control the original image index coordinates, rather than the world indices.
This is not possible now. This would need the image navigator to check if the renderer
axes are flipped compared to the corresponding reference geometry axes, and invert
the directions if and only if they are.