Issue #4837: DnD Display relies on MITK when initialising renderers
Now that MITK can initialise renderers quasi correctly from any image geometry
(flipped, permuted or rotated axes), we can finally remove a whole lot of
complicated code for creating these geometries by hand in niftkMultiWindowWidget.

Actually, this code *has* to be removed now because it initialised the slice
navigation controllers using the 'Original' view direction, so that the geometries
can be created beforehand and pushed down the throught of the SNC, without that
it creates a new sliced geometry from the input geometry.

This worked well, but in this way the input geometry (at other places also called
'reference' geometry) of the renderers were the geometries that we created
by hand, and not the one that we used as input to create the input geometries for
the renderers.

So, when the image navigator checks if the direction of the SNC is inverted
against the direction of the corresponding axis of the reference geometry, it
finds one of the three geometries that we created for the renderers as input,
but not the actual image geometry.

One way to tackle with this would be to check in the image navigator if the
input geometry of the SNC has a reference geometry, and if it does, it could
use that.

However, creating the input geometries manually is not necessary any more,
and hopefully the fix that allows this will get merged to the MITK upstream,
soon.

So, both for fixing this and for the sake of simplicity, we remove this big
chunk of code from the DnD Display, finally.