Issue #4837: Equality check when setting time geometry for single viewer
This was probably omitted to allow reinitialising the viewer with the same
geometry.

However, the rigorous unit tests fail because setting the same geometry again
raises signals even when the state of the viewer does not change. The extra
events are not a severe problem, as the state is consistent, but it can make
other objects (observers) update themselves even if they would not need to.

So, we might revert this later if it is found out why the events are sent out.
This would let the user reinitialise the viewer by dragging the same image
on the viewer again.