Issue #4837: Initialise viewer with input geometry of renderers
It is not supported any more to initialise the dnd display viewer with a
geometry that comes directly from the renderer. These geometries are a bit
messed up at the moment, e.g. they are shifted with half a voxel relative
to the original, 'reference' geometry that was used to initialise them.

Instead, you should take the 'input' geometry of the slice navigation
controller of the renderer, and use that to initialise the dnd display
viewer.