Issue #4590: Fix build directory names
Not just the external project directory but also the superbuild directory must be
different for 32bit and 64bit windows builds, if they are on the same machine.

The use of the external project directory has been removed from release jobs as
they must not use the same external project directory that is used by the continuous
jobs. Release builds are always full builds, because they build the external
projects with different options. (E.g. Python 'on'.)