Issue #4590: Enable 32bit windows jobs
Note:
Until now these jobs were commented out and the global before_script was overwritten
in them so that the external project folder is different than the one for the 64bit
builds.

With this commit I uncommented the 32bit jobs and removed the '-x86' suffix from the
external project build folder name, because it is likely that with the suffix some
paths would be too long for MITK or ITK.

I assume that the 32bit and 64bit runners are running on either different machines
or - if on the same machine - they are different runner instances.

You can run multiple runners on the same machine if you put multiple '[[runners]]'
section in the connfiguration file of the runner ('config.toml').