Issue #4590: Use different superbuild dir for continuous and release builds
Although the continuous and release builds build the external projects with
different options (e.g. with Python integration off or on, or in debug or
release mode), there is no conflict in terms of the external project directory
because the release builds do a complete build and they do not use the EP
directory that is shared by the continuous builds.

However, the superbuild directory is named after the commit hash, and it is
the same for tags. So, if you push the master then tag it and push the tag,
the continuous builds and the release builds will be made under the same
path.

This is usually not a problem, because you push the master, wait for the
pipeline to finish, and if it is 'green', you can tag a push the tag. At the
end of the pipeline the superbuild folder is deleted, so the release build
can create a new directory with the same name.

There can be clash only if you push the tag before the pipeline for the same
commit has finished or if some of the continuous clean-up jobs are commented
out for testing.

And of course, it is an issue only if the release and continuous builds are
run on the same computer by the same runner instance.

[ci skip]