Issue #4876: Ignore width of file name labels
Too long file names broke the GUI. They made some widgets really wide,
and a scrollbar was put on the bottom so that you can scroll the widget.

With this change the width of the labels with file names are ignored.
If the text is longer, the end is chopped off, but you can always see
the full name in the tool tip.