Issue #4904: Store segmentation node in tool manager working data
The general segmentor put the segmentation node in the working data vector
of the tool manager as the 0th element. The morphological segmentor did not
do this, since its only tool (paintbrush tool) did not use it.

However, the image selector widget that was used in both segmentor views,
assumed that the segmentation is the 0th working data.

This change adds the segmentation node in the working data vector also
for the morphological segmentor view, so that the image selector widget
can work correctly.