Issue #4957: Always add children when parent dragged
The children were added only when they were globally visible. This caused
that if they were not visible when they had been dragged, it was not possible
to 'switch them on' by the check box in the data manager. You had to select
them together with the parent (by shift or ctrl or cmd key) and drag them
together in the viewer.

This was contra-intuitive. The children are now always added to the viewer
when the parent is dragged, and their initial visibility will be their global
visibility.