Issue #4962: Fix crash because of registering focus listeners in wrong order
The crash happened if the DnD display had two viewers, you opened a reference image and a segmentation
in both viewers and restarted the segmentation with the irregular editor.

The crash did not happen always, and I only saw it on Linux, not on Mac.

I found that the crash happens if the irregular editor is created earlier than the DnD display.

The cause of the crash was that both the irregular editor (more precisely niftk::BaseController),
and the DnD display registered their listener to the focus manager's focus events. However, if the
irregular editor was created first, it registered its listener first, and accordingly it was notified
of the focus events earlier then the DnD display itself (niftk::MultiViewerWidget).

So, the irregular editor got to know about the focus change *before* the multi viewer widget, which
could not update its state. When the irregular editor got to know about the focus change, the multi
viewer still stored the previous viewer as the 'selected' viewer, altough the focus was already on
the new one.

The solution is that the multi viewer does not use the focus listener at all, but it relies on
the Qt signal 'WindowSelected()'. The signal is sent out *before* the focus event, and in this
way we can be sure that it is processed in time.