Issue #4962: Emit Qt signal after changing the focus through the focus manager
This basically undoes what was done in a previous commit. But if the Qt signal
was emitted before notifying the focus manager, many unit tests failed because
the selected window did not match the focused window when the Qt signal was
caught. This means that the viewer is in an inconsistent state. (Or the state
of the viewer is not consistent with the state of the focus manager.)

So, I had to revert this, and have to find a different fix.