Issue #4962: Do not rely on MultiViewer::GetSelectedRenderWindow() in focus event listener
The order in which the MultiViewer and the segmentor views are notified about focus events,
depends on the order in which they are created. If the segmentor view is created first,
it will get notified about the focus change events before the multi viewer. Therefore,
when the segmentor knows about the focus change, the multi viewer could not update his
state to store which is the new 'selected' viewer.

This commit changes the logic of a focus listener in BaseController so that it does not
call the GetSelectedViewer function of the multi viewer.