Issue #4969: RGB pixel type support for looking for unenclosed seeds
Work in progress.

Compilation fails because niftk::GeneralSegmentorPipeline declares
members with this type:

    itk::CastImageFilter<GreyScaleImageType, SegmentationImageType>

But the RGB pixel type cannot be casted to a scalar type.

As I see, the cast image filters should not be used at all, they are
used only to create an initial binary mask with the same geometry as the
reference image.

[ci skip]