Issue #4997: Always set renderer specific visibility in multi window widget
niftk::MultiWindowWidget::SetVisibility(QmitkRenderWindow*, mitk::DataNode*, bool)
set the renderer specific visibilities only if the new visibility was different than
the old renderer specific visibility (so far so good) OR if there was no renderer
specific visibility then the global visibility.

It was probably not meant to work like this, but this call:

    node->IsVisible(renderer)

results the global visibility if there is no renderer specific one.

As a result, in many cases the renderer specific visibility was not set by
this function because it was the same as the global visibility.

The multi viewer visibility manager needs that each data node has global and
renderer specific visibilities for each renderer of each viewer.