Issue #5035: Use ITK transform functions to convert coordinates
The index coordinates were converted to world space (mm) manually,
without using the ITK transformation function that's specifically for
this ('TransformContinuousIndexToPhysicalPoint'). This made the code
unnecessarily complicated and more difficult to read, and on top of
that, the rotation matrix was not applied, and the result was wrong for
oblique images. Permuted axes and spacing was considered, though.

Using the transformation matrix for converting between index and world
coordinates simplifies the code and makes it work for arbitrary images.