Issue #5035: Convert contour points to index coordinates for comparison
The contours and contour sets store mm positions now, and the geometry
of these contours and contour sets is a default geometry with unity
rotation matrix and unit spacings. Although the contours are rendered
correctly, this is wrong because it does not make it possible to recover
the original indices. Ideally, the contours should store the index
coordinates and the geometry should be the same as that of the reference
image or the mask.

It would be too much work to correct this, especially at this stage of
the development. Also, there is only one function where you need the
original indices. So, as a workaround we pass the geometry as an
additional argument to this function. The geometry is used to convert
the world coordinates (that are stored as contour points) back to index
coordinates.