Issue #5047: Confirmation dialog before deleting perspective
Another change is that the manually created perspectives are now added
to the end of the Open Perspective menu, and not inserted in lexicographic
order. This is better because when the application is restarted, the pre-defined
perspectives will be loaded first, and they will appear on the top of the menu.
That is, the perspectives are not going to be in lexicographic order, the manual
perspectives will be at the end.