Issue #5079: Declare more image arguments as const
Declaring image arguments as const not only ensures static compiler
checks but the read/write lock mechanism is also based on the constness
of the image arguments. The ITK access functions of MITK apply write
lock for the image which are not passed as const. Simoultaneous accesses
to an image (when at least one of the locks is for writing) will cause
error at run-time.

This commit adds the 'const' qualifier to many places.

[ci skip]