Issue #5079: Transfer ownership of ITK image to MITK
The application crashed when the 'retain marks' option was switched on.
This happened because the result of the 'retain marks' calculation was
an ITK image that has been converted to an MITK image by mitk::ImportItkImage().
After the conversion the ITK smart pointers have been released.

However, mitk::ImportItkImage() does not transfer the ownership of the buffer to
the MITK image. You can use the MITK image only until the ITK image lives. In this
case, however, the MITK image was put in a data node, and the ITK image was released.

The application crashed at the first time when the MITK image was read.

In principle, mitk::GrabItkImageMemory() should transfer the ownership of the buffer
to the MITK image, exactly to prevent this kind of errors. For some reason, it did
not help here, so I used mitk::CastToMitkImage().