Issue #5122: Code tidy up
Unused bool return value removed. I removed the nested if-s and return
early from functions if the data is not valid. Some people argue that
it is not nice to have several exit points in a function. It is a valid
point, but I do not like nested if-s at so many levels. It is hard to
keep it in mind which level I'm at, and here they were not really for
the 'logic', but just to check data before running the actual
processing.

I also removed the extra 'std::endl' from the MITK_ERROR macros because
they add the end line by default.