Issue #5127: Remove code for switching window layout by double click
The interaction was already disabled, but the code still existed in the
code base. Sometimes it caused error when an error dialog popped up
because saying it is not possible to pop values from an empty deque.

Since the interaction is disabled anyway, it is not worth investigating
what causes the error, it is easier to remove the obsolete code
completely.

The interaction had to be disabled for the request of Emily Manning
because it was easy to double click accidentally during segmentation.