Issue #5134: Data Manager context menu to load data into viewer
This is possible through drag and drop, but the user can close the editor and there is
no way to reopen it. Global reinit reopened the editor, but this action has been removed
by a previous commit on this branch because it messed up the geometry of the DnD Display
viewers.