Issue #5170: Remove state variable that was independent of segmentation
So that we can switch across multiple segmentations, the state of the
segmentation process must be stored in the segmentation node, or at
least it should be associated to it in some way.

There was a "m_TabIndex" member variable that was not associated to any
segmentation node, but it stored a part of the state of the GUI. This
variable was not updated when the selected data node (selected
segmentation) has changed. This could potentially cause problems. E.g.
if one segmentation was at stage (tab index) 1 and the other was at
stage 2 and you switched from the first segmentation to the seconde one,
the segmentor might eventually think that you changed stage, although
you did not. Or something like that.

To avoid such potential errors, the internal state of the segmentor has
been eliminated. In this case, when you change tab, the previous tab
index is taken from a property of the segmentation node.