Issue #5162: Correct index diffence between working data nodes and editing flags
The working data nodes used by the paintbrush tool are indexed from 1 to 4, but the vector
of the editing flags used by the morphological segmentor pipeline is indexed from 0 to 3.

In the past they used to match but now the segmentation image is also stored in the working
data node vector with index 0. This was not like that in the past, so the two indices used
to match. However, so that the image selector widget works correctly and common code for
the morphological and irregular segmentor can be pulled up to a common base class, it
was necessary to make this change.

This commit applies the 1 difference when indexing the two different vectors.

The rest is just tidying.