Issue #5163: Hide eraser cursor immediately after interaction
After you finished adding or subtracting voxel by the paintbrush tool,
the eraser cursor stayed on the screen for a couple of seconds, and so,
it looked as if the application was not responsive.

Indeed, the case was only that the cursor was removed after the connected
component analysis has finished, not immediately after the mouse button has
been released. The other reason was that the screen was not updated immediately,
just an update has been requested (but not forced).