Issue #5169: Do not 'optimise' seeds when applying threshold
There is an 'optimised' flag that determines if seeds should be regenerated for
closed regions. The flag is set to 'true' when thresholding is applied. (This
happens if you press the 'Apply' or 'Clean' button or change slice.

This commit changes the 'optimised' flag to false to leave the seeds as they are.