Issue #5170: Fix for uninitialised local variables.
FindFirstParentImage did not initialise a bool variable for the property to retrieve.
This could be a problem if the property is not present in the data node. The return
value of the GetBoolProperty() function was not checked.

At other places the bool variables were initialised, although the return value
of the GetBoolProperty function was checked.

Smart pointers are initialised with nullptr, so the explicit initialisation was
unnecessary.