Issue #5170: Reinitialise segmentor controller with data manager after focus change
The MIDAS segmentors assume that the reference image that is being segmented
has the same geometry as that was used to initialise the selected render windows
in which you see the images and do interactions.

One of the last commits introduced a check to disable the segmentor tools
and disallow creating new segmentations if the geometry of the reference image
does not match the input geometry of the renderer. This can prevent a misuse
when you are editing a wrong segmentation, not the one you think.

There was a little problem with the implementation, though. The data manager
selection change event is fired earlier than the focus change event. So, when
the data node selection has changed and a potential reference image is selected,
it will not have the same geometry as the focused viewer, because the focus has
not transferred yet.

As a solution, you have to simulate a data node selection change event after
the focus has transferred. The geometries will match this time.

A bug has been discovered as well. Although, the BaseController class had
a OnFocusChanged() function that derived classes could override, it was never
called. This is fixed.

The commit also simplifies the API for determining if a node can serve as a
reference or segmentation image.